gradient work
Added some css, mostly worked on gradient stuff. Moved all noise calculations to a seperate thread. Updates to the background happen whenever we get a new set of data from the web worker. To avoid jitteryness, we can just slow the z_off. Basically, as long as we are using the cpu to calc the noise per pixel for the entire screen, it's going to be slow; there's nothing that can be done about that. The other options were to do the calculations via webGL and display it all that way, but that would involve moving into GLSL (maybe at some point), or to fake it with some blurry balls. Also, main.js needs to be cleaned up., The perlin noise function still exists in the files, but was removed from index.html since it's not currently used anywhere except the noiseWorker, which has it's own global space.